+++ frontend/src/features/Reporting/components/VesselReportingList/Current/ConfirmDeletionModal.tsx
