+++ frontend/src/features/Reporting/components/VesselReportingList/Current/utils.ts
