+++ frontend/src/features/Reporting/components/VesselReportingList/Summary/InfractionSuspicionSummary.tsx
