+++ frontend/src/features/Reporting/components/VesselReportingList/Summary/index.tsx
@@ -1,4 +1,4 @@
-import { InfractionSuspicionSummary } from '@features/Reporting/components/CardReportingList/Summary/InfractionSuspicionSummary'
+import { InfractionSuspicionSummary } from '@features/Reporting/components/VesselReportingList/Summary/InfractionSuspicionSummary'
 import { Header, Zone } from '@features/Vessel/components/VesselSidebar/common_styles/common.style'
 import { useGetVesselReportingsByVesselIdentityQuery } from '@features/Vessel/vesselApi'
 import { useMainAppSelector } from '@hooks/useMainAppSelector'
