+++ frontend/src/features/Reporting/components/VesselReportingList/__tests__/ReportingCard.test.tsx
@@ -1,5 +1,5 @@
 import { Seafront } from '@constants/seafront'
-import { ReportingCard } from '@features/Reporting/components/CardReportingList/ReportingCard'
+import { ReportingCard } from '@features/Reporting/components/VesselReportingList/ReportingCard'
 import { ReportingType } from '@features/Reporting/types'
 import { afterAll, describe, expect, it } from '@jest/globals'
 import { THEME, ThemeProvider } from '@mtes-mct/monitor-ui'
